refactor: Rename `lib/plugins/aws/package/lib/saveCompiledTemplate.js`
