+++ lib/plugins/aws/package/index.js
@@ -6,7 +6,7 @@ const stripNullPropsFromTemplateResources = require('./lib/strip-null-props-from
 const generateArtifactDirectoryName = require('./lib/generate-artifact-directory-name');
 const generateCoreTemplate = require('./lib/generate-core-template');
 const saveServiceState = require('./lib/save-service-state');
-const saveCompiledTemplate = require('./lib/saveCompiledTemplate');
+const saveCompiledTemplate = require('./lib/save-compiled-template');
 const mergeIamTemplates = require('./lib/merge-iam-templates');
 const addExportNameForOutputs = require('./lib/add-export-name-for-outputs');
 const { log, style, progress } = require('@serverless/utils/log');
