+++ lib/plugins/aws/package/lib/save-compiled-template.js
