+++ test/unit/lib/plugins/aws/package/lib/save-compiled-template.test.js
